client/server and hooks working w/ brpc
So I'm not super happy with the brpc API. I mean it's fine, I think.
It's not super clunky, but what I think it might be is that there is
one API for a few different use cases that could be 2-3 different APIs
that would be better for each. The hard part is that it might increase
the amount of code 2-3x without a lot of abstraction. But maybe that's
fine.